Improve pole placement
